requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
crypto==1.4.1
Flask==2.0.2
Flask-PyMongo==2.3.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Naked==0.1.31
passlib==1.7.4
PyJWT==2.3.0
pymongo==3.12.1
python-dotenv==0.19.2
PyYAML==6.0
requests==2.26.0
shellescape==3.8.1
urllib3==1.26.7
Werkzeug==2.0.2
